feat: use files for aapt command-line arguments, reorganize default paths

+++ xbuild.py
@@ -4,7 +4,18 @@ from R_java import updateR
 from dependency import ProjectDependency,Dependency
 import re
 import zipfile
-
+class BuildPaths:
+    def __init__(self,conf=None):
+        super().__init__()
+        if not conf:
+            conf = {}
+        self.java = '/'+conf.get("java_path","src/main/java")
+        self.aidl = '/'+conf.get("aidl_path","src/main/aidl")
+        self.libs = '/'+conf.get("libs_path","libs")
+        self.res = '/'+conf.get("res_path","src/main/res")
+        self.manifest = '/'+conf.get("manifest","src/main/AndroidManifest.xml")
+        self.build = '/'+conf.get("build_path","xbuild")
+        self.key = '/'+conf.get("key_path","main.key")
 class BuildModes:
     CLEAN = 0
     CLEAN_KEEP_JARDEX = 1
@@ -14,49 +25,43 @@ class BuildModes:
 def setupPaths(lib):
     p =os.path
     path=lib.path
-    lib.java_path = f"{path}/src/main/java"
-    lib.manifest = f"{path}/src/main/AndroidManifest.xml"
-    lib.res_path = f"{path}/src/main/res"
-    lib.build_path = f"{path}/xbuild"
+    lib.java_path = path + PATHS.java
+    lib.manifest = path + PATHS.manifest
+    lib.res_path = path + PATHS.res
+    lib.build_path = path + PATHS.build
+    lib.aidl_path = path + PATHS.aidl
     lib.dex_path = lib.build_path+"/bin/classes.dex"
-    lib.aidl_path = f"{path}/src/main/aidl"
     if not os.path.exists(lib.aidl_path):
         lib.aidl_path = None
     if USE_AAPT2:
         flat_path = lib.build_path+"/resources/"
         lib.flat_res_path = flat_path
         lib.flat_res_tree = readFlatResTree(lib,wrap(lib.flat_res_path))    
-        if lib.res_path:
-            if not p.exists(flat_path):
-                os.makedirs(flat_path)
     else:
         lib.res_tree = readResTree(lib,wrap(lib.res_path))
 
 def setupAarPaths(lib):
     zipper = zipfile.ZipFile(lib.path)
     lib.java_path = None
-    lib.build_path = f"{lib.projectPath}/xbuild"
-    lib.manifest = lib.projectPath+"/AndroidManifest.xml"
+    lib.build_path = lib.projectPath + PATHS.build
+    lib.manifest = lib.projectPath+PATHS.manifest
     lib.dex_path=None 
-    lib.res_path = lib.projectPath+"/res"
+    lib.res_path = lib.projectPath+PATHS.res;
+    lib.aidl_path = lib.projectPath+PATHS.aidl;
+    libPath = lib.projectPath+PATHS.libs;
     zipper.extractall(lib.projectPath)
     zipper.close()
     if not path.exists(lib.res_path):
         lib.res_path=None
     lib.jar_files = glob(lib.projectPath,"*.jar")
-    lib.aidl_path = lib.projectPath+"/aidl"
     if not path.exists(lib.aidl_path):
         lib.aidl_path = None
-    libPath = lib.projectPath+"/libs"
     if path.exists(libPath):
         lib.jar_files.extend(glob(libPath,"*.jar"))
     if USE_AAPT2:
         flat_path = lib.build_path+"/resources/"
         lib.flat_res_path = flat_path
-        lib.flat_res_tree = readFlatResTree(lib,wrap(lib.flat_res_path))    
-        if lib.res_path:
-            if not path.exists(flat_path):
-                os.makedirs(flat_path)
+        lib.flat_res_tree = readFlatResTree(lib,wrap(lib.flat_res_path))
     else:
         lib.res_tree = readResTree(lib,wrap(lib.res_path))
  
@@ -64,7 +69,7 @@ def setupJarPaths(lib):
     lib.manifest = None
     lib.res_path = None
     lib.jar_files = [lib.path]
-    lib.build_path = f"{lib.projectPath}/xbuild"
+    lib.build_path = lib.projectPath + PATHS.build
     lib.res_tree=[]
     lib.java_path = None
     lib.dex_path=None
@@ -74,16 +79,10 @@ def readResTree(lib,tree):
     for i in lib.dependencies:
         if i.type == "unresolved":
             continue
-        #if i.res_path:
-        #    if i.res_path in tree:
-        #        tree.remove(i.res_path)
-        #    tree.append(i.res_path)
         for j in i.res_tree:
             if j in tree:
                 tree.remove(j)
             tree.append(j)
-            
-        #readResTree(i,tree)
     assert(checkSet(tree))
     return tree
 
@@ -93,16 +92,10 @@ def readFlatResTree(lib,tree):
             continue
         if not i.res_path:
             continue
-        #if i.res_path:
-        #    if i.res_path in tree:
-        #        tree.remove(i.res_path)
-        #    tree.append(i.res_path)
         for j in i.flat_res_tree:
             if j in tree:
                 tree.remove(j)
             tree.append(j)
-            
-        #readResTree(i,tree)
     assert(checkSet(tree))
     return tree
 
@@ -110,9 +103,6 @@ def readDepTree(lib,tree):
     for i in lib.dependencies:
         if i.type == "unresolved":
             continue
-        #    if i.res_path in tree:
-        #        tree.remove(i.res_path)
-        #    tree.append(i.res_path)
         if i in tree:
             tree.remove(i)
         tree.append(i)
@@ -121,26 +111,30 @@ def readDepTree(lib,tree):
     return tree
 
 def readAidlTree(lib,lists):
-        if lib.aidl_path:
-            if lib.aidl_path in lists:
-                return lists
-            lists.append(lib.aidl_path)
+    if lib.aidl_path:
+        if lib.aidl_path in lists:
+            return lists
+        lists.append(lib.aidl_path)
         for i in lib.dependencies:
             readAidlTree(i,lists)
-        return lists
+    return lists
     
 def buildResources(lib,updateChildren=False):
     info("aapt compiling "+lib.name)
     manifest = lib.manifest
     gen_path = lib.build_path+"/gen"
+    if not path.exists(gen_path):
+        os.makedirs(gen_path)
+    
     assert lib.res_path
     a = None
     if not USE_AAPT2:
         res_path = lib.res_tree
-        if not path.exists(gen_path):
-            os.makedirs(gen_path)
         a = aapt_package(manifest,res_path,gen_path,lib.type=="project",lib.getOptions())
     else:
+        if not path.exists(lib.flat_res_path):
+            os.makedirs(lib.flat_res_path)
+        #TODO UNCHANGED
         a = aapt2_compile(lib.res_path,lib.flat_res_path)
         if a:
             a = aapt2_link(lib.flat_res_tree,gen_path,manifest,lib.getOptions())
@@ -160,6 +154,12 @@ def build(lib,mode=BuildModes.CLEAN_KEEP_JARDEX):
     obj_path = build_path+"/obj"
     bin_path = build_path+"/bin"
     gen_path = build_path+"/gen"
+    
+    if mode<BuildModes.UNCHANGED:
+        if path.exists(obj_path):
+            recursiveDelete(obj_path)
+        if path.exists(bin_path):
+            recursiveDelete(bin_path)
     java_files = []
     java_path = wrap(lib.java_path)
     for i in java_path:
@@ -209,6 +209,7 @@ def build(lib,mode=BuildModes.CLEAN_KEEP_JARDEX):
             raise Exception("aidl_failed")
     java_files.extend(glob(gen_path,"**.java"))
     if len(java_files) > 0:
+        #TODO UNCHANGED
         a = ecj_compile(java_files,classpath,source_path,obj_path)
         if not a:
             raise Exception("ecj failed")
@@ -220,7 +221,9 @@ def buildDependencyTree(head,mode = BuildModes.CLEAN_KEEP_JARDEX):
     if head:
         head.collect()
     libraries = readDepTree(head,wrap(head))[-1::-1]
-    print("\n-".join(map(lambda a:a.name,libraries)))
+    debug("\n-".join(map(lambda a:a.name,libraries)))
+    
+    #Setup paths and build resources
     for dep_lib in libraries:
         if dep_lib.type == "project":
             setupPaths(dep_lib)
@@ -257,8 +260,8 @@ def buildApp(app,app_name,mode=BuildModes.CLEAN_KEEP_JARDEX):
     dexList = []
     for i in Dependency.globalDeps:
         j = Dependency.globalDeps[i]
-        if j.dex_path:
-            dexList.append(j.dex_path)
+        #if path.exists(j.dex_path):
+            #dexList.append(j.dex_path)
         if path.exists(j.build_path+"/dexLibs"):
             dexList.extend(glob(j.build_path+"/dexLibs","*.dex.jar"))
         assert checkSet(dexList), j.name
@@ -268,33 +271,73 @@ def buildApp(app,app_name,mode=BuildModes.CLEAN_KEEP_JARDEX):
         os.makedirs(mergedPath)
     binPath = app.build_path+"/bin"
     info("merging dex")
+    status = False
     if mode is BuildModes.FAST and path.exists(mergedPath+"/classes.dex"):
-        dx_dex(app.getObjs(False),app.dex_path)
-        dx_merge(mergedPath+"/classes.dex",app.dex_path,mergedPath+"/classes.dex")
+        status = dx_dex(app.getObjs(False),app.dex_path)
+        if status:
+            status = dx_merge(mergedPath+"/classes.dex",app.dex_path,mergedPath+"/classes.dex")
+        else: return error('Dexing Failed')
     else:
-        dx_dex(app.getObjs(True),app.dex_path)
-        dx_merge(mergedPath+"/classes.dex",*dexList)
+        status = dx_dex(app.getObjs(True),app.dex_path)
+        if status:
+            status = dx_merge(mergedPath+"/classes.dex",app.dex_path,*dexList)
+        else: return error('Dexing Failed')
+    if not status:
+        return error("Merge Failed")
     apk_unaligned = binPath+f"/{app_name}-unalingned.apk"
     if path.exists(apk_unaligned):
         os.remove(apk_unaligned)
     apk_file = binPath+f"/{app_name}.apk"
-    os.system(f"zipmerge {apk_unaligned} "+" ".join(filter(lambda a:a.endswith("jar"),dexList)))
-    packager = (USE_AAPT2 and aapt2_package_res) or package_res
-    a = packager(app.manifest,app.res_tree,mergedPath,apk_unaligned,app.getOptions())
+    if len(dexList)>0:
+        os.system(f"zipmerge {apk_unaligned} "+" ".join(filter(lambda a:a.endswith("jar"),dexList)))
+    if USE_AAPT2:
+        a = aapt2_package_res(app.manifest,app.flat_res_tree,mergedPath,apk_unaligned,app.getOptions())
+    else:
+        package_res(app.manifest,app.res_tree,mergedPath,apk_unaligned,app.getOptions())
     if not a:
         return "packaging failed"
-    a = sign_align(apk_unaligned,SIGNING_KEY_PATH,apk_file)
+    a = sign_align(apk_unaligned,app.path+PATHS.key,apk_file)
     os.remove(apk_unaligned)
     if not a:
         return "signing/zipaligning failed"
     install(apk_file);
     return "Succeeded"
 
-def main(path="."):
+def main(path=".",mode=BuildModes.CLEAN):
     app = ProjectDependency("main",path)
-    buildApp(app,"app",BuildModes.FAST)
+    buildApp(app,"app",mode)
 def mainLibrary(path="."):
     pass
 if __name__ == "__main__":
-    main(FIRST)
+    if HELP:
+        print('\n    '.join((
+            'xbuild.py [-h][-d][-q][-a][-c configfile] [-m CLEAN|FAST|UNCHANGED|SOURCE] [dir]',
+                'Xbuild - Utility to build Gradle android projects',
+                'h help',
+                'd debug',
+                'q quiet',
+                'c configFile',
+                'm BuildMode',
+                'l List Dependencies'
+                'a Use aapt2 instead of aapt',
+                'SOURCE rebuild all source files',
+                'CLEAN rebuild both source files and libraries',
+                'FAST rebuild main project source files',
+                'UNCHANGED unimplemented, works as fast'
+                )))
+    elif LIST_DEPS:
+        dep = ProjectDependency("main",TARGET)
+        dep.collect()
+        print(dep)
+    else:
+        PATHS=BuildPaths(CONFIG)
+        mode = BuildModes.CLEAN;
+        if MODE:
+            if MODE == 'FAST':
+                mode = BuildModes.FAST
+            elif MODE == 'UNCHANGED':
+                mode = BuildModes.UNCHANGED
+            elif MODE == 'SOURCE':
+                mode = BuildModes.CLEAN_KEEP_JARDEX
+        main(TARGET,mode)
